+++ csharp/sdk/4.0/methods.cs
@@ -286,8 +286,8 @@ namespace Looker.SDK.API40
   /// ### Present client credentials to obtain an authorization token
   ///
   /// Looker API implements the OAuth2 [Resource Owner Password Credentials Grant](https://cloud.google.com/looker/docs/r/api/outh2_resource_owner_pc) pattern.
-  /// The client credentials required for this login must be obtained by creating an API3 key on a user account
-  /// in the Looker Admin console. The API3 key consists of a public `client_id` and a private `client_secret`.
+  /// The client credentials required for this login must be obtained by creating an API key on a user account
+  /// in the Looker Admin console. The API key consists of a public `client_id` and a private `client_secret`.
   ///
   /// The access token returned by `login` must be used in the HTTP Authorization header of subsequent
   /// API requests, like this:
@@ -310,14 +310,14 @@ namespace Looker.SDK.API40
   /// ### Best Practice:
   /// Always pass credentials in body params. Pass credentials in URL query params **only** when you cannot pass body params due to application, tool, or other limitations.
   ///
-  /// For more information and detailed examples of Looker API authorization, see [How to Authenticate to Looker API3](https://github.com/looker/looker-sdk-ruby/blob/master/authentication.md).
+  /// For more information and detailed examples of Looker API authorization, see [How to Authenticate to Looker API](https://github.com/looker/looker-sdk-ruby/blob/master/authentication.md).
   ///
   /// POST /login -> AccessToken
   ///
   /// <returns><c>AccessToken</c> Access token with metadata. (application/json)</returns>
   ///
-  /// <param name="client_id">client_id part of API3 Key.</param>
-  /// <param name="client_secret">client_secret part of API3 Key.</param>
+  /// <param name="client_id">client_id part of API Key.</param>
+  /// <param name="client_secret">client_secret part of API Key.</param>
   public async Task<SdkResponse<AccessToken, Exception>> login(
     string? client_id = null,
     string? client_secret = null,
@@ -3246,6 +3246,7 @@ namespace Looker.SDK.API40
   /// <param name="type">Either dashboard or look</param>
   /// <param name="resource_id">ID of the dashboard or look to render</param>
   /// <param name="reload">Whether or not to refresh the rendered image with the latest content</param>
+  /// <param name="theme">Light or dark background. Default is "light"</param>
   /// <param name="format">A value of png produces a thumbnail in PNG format instead of SVG (default)</param>
   /// <param name="width">The width of the image if format is supplied</param>
   /// <param name="height">The height of the image if format is supplied</param>
@@ -3253,6 +3254,7 @@ namespace Looker.SDK.API40
     string type,
     string resource_id,
     string? reload = null,
+    string? theme = null,
     string? format = null,
     long? width = null,
     long? height = null,
@@ -3262,6 +3264,7 @@ namespace Looker.SDK.API40
       resource_id = SdkUtils.EncodeParam(resource_id);
     return await AuthRequest<TSuccess, Exception>(HttpMethod.Get, $"/content_thumbnail/{type}/{resource_id}", new Values {
       { "reload", reload },
+      { "theme", theme },
       { "format", format },
       { "width", width },
       { "height", height }},null,options);
@@ -3437,7 +3440,8 @@ namespace Looker.SDK.API40
 
   /// ### Search Dashboards
   ///
-  /// Returns an **array of dashboard objects** that match the specified search criteria.
+  /// Returns an array of **user-defined dashboard** objects that match the specified search criteria.
+  /// Note, [search_dashboards()](#!/Dashboard/search_dashboards) does not return LookML dashboard objects.
   ///
   /// If multiple search params are given and `filter_or` is FALSE or not specified,
   /// search params are combined in a logical AND operation.
@@ -9542,14 +9546,14 @@ namespace Looker.SDK.API40
     return await AuthRequest<string, Exception>(HttpMethod.Delete, $"/users/{user_id}/credentials_oidc", null,null,options);
   }
 
-  /// ### API 3 login information for the specified user. This is for the newer API keys that can be added for any user.
+  /// ### API login information for the specified user. This is for the newer API keys that can be added for any user.
   ///
   /// GET /users/{user_id}/credentials_api3/{credentials_api3_id} -> CredentialsApi3
   ///
-  /// <returns><c>CredentialsApi3</c> API 3 Credential (application/json)</returns>
+  /// <returns><c>CredentialsApi3</c> API Credential (application/json)</returns>
   ///
   /// <param name="user_id">Id of user</param>
-  /// <param name="credentials_api3_id">Id of API 3 Credential</param>
+  /// <param name="credentials_api3_id">Id of API Credential</param>
   /// <param name="fields">Requested fields.</param>
   public async Task<SdkResponse<CredentialsApi3, Exception>> user_credentials_api3(
     string user_id,
@@ -9563,14 +9567,14 @@ namespace Looker.SDK.API40
       { "fields", fields }},null,options);
   }
 
-  /// ### API 3 login information for the specified user. This is for the newer API keys that can be added for any user.
+  /// ### API login information for the specified user. This is for the newer API keys that can be added for any user.
   ///
   /// DELETE /users/{user_id}/credentials_api3/{credentials_api3_id} -> string
   ///
   /// <returns><c>string</c> Successfully deleted. (application/json)</returns>
   ///
   /// <param name="user_id">Id of user</param>
-  /// <param name="credentials_api3_id">Id of API 3 Credential</param>
+  /// <param name="credentials_api3_id">Id of API Credential</param>
   public async Task<SdkResponse<string, Exception>> delete_user_credentials_api3(
     string user_id,
     string credentials_api3_id,
@@ -9581,11 +9585,11 @@ namespace Looker.SDK.API40
     return await AuthRequest<string, Exception>(HttpMethod.Delete, $"/users/{user_id}/credentials_api3/{credentials_api3_id}", null,null,options);
   }
 
-  /// ### API 3 login information for the specified user. This is for the newer API keys that can be added for any user.
+  /// ### API login information for the specified user. This is for the newer API keys that can be added for any user.
   ///
   /// GET /users/{user_id}/credentials_api3 -> CredentialsApi3[]
   ///
-  /// <returns><c>CredentialsApi3[]</c> API 3 Credential (application/json)</returns>
+  /// <returns><c>CredentialsApi3[]</c> API Credential (application/json)</returns>
   ///
   /// <param name="user_id">Id of user</param>
   /// <param name="fields">Requested fields.</param>
@@ -9599,11 +9603,11 @@ namespace Looker.SDK.API40
       { "fields", fields }},null,options);
   }
 
-  /// ### API 3 login information for the specified user. This is for the newer API keys that can be added for any user.
+  /// ### API login information for the specified user. This is for the newer API keys that can be added for any user.
   ///
   /// POST /users/{user_id}/credentials_api3 -> CreateCredentialsApi3
   ///
-  /// <returns><c>CreateCredentialsApi3</c> API 3 Credential (application/json)</returns>
+  /// <returns><c>CreateCredentialsApi3</c> API Credential (application/json)</returns>
   ///
   /// <param name="user_id">Id of user</param>
   /// <param name="fields">Requested fields.</param>
@@ -10179,7 +10183,7 @@ namespace Looker.SDK.API40
   ///
   /// The dev workspace is NOT unique to an API session. Two applications accessing the Looker API using
   /// the same user account will see the same files in the dev workspace. To avoid collisions between
-  /// API clients it's best to have each client login with API3 credentials for a different user account.
+  /// API clients it's best to have each client login with API credentials for a different user account.
   ///
   /// Changes made to files in a dev workspace are persistent across API sessions. It's a good
   /// idea to commit any changes you've made to the git repository, but not strictly required. Your modified files