+++ packages/embed-services/src/test-utils/index.ts
@@ -23,4 +23,4 @@
  SOFTWARE.
 
  */
-export { OAuthConfigForm } from './OAuthConfigForm'
+export * from './utils'