+++ packages/extension-utils/src/OAuthForm/src/components/OAuthForm.tsx
@@ -24,8 +24,8 @@
 
  */
 import type { BaseSyntheticEvent, Dispatch, FormEvent } from 'react'
-import React, { useState, useEffect } from 'react'
-import type { ValidationMessages, MessageBarIntent } from '@looker/components'
+import React, { useEffect } from 'react'
+import type { ValidationMessages } from '@looker/components'
 import {
   Button,
   ButtonTransparent,
@@ -44,32 +44,14 @@ import {
 import { CodeCopy } from '@looker/code-editor'
 import type { ILookerVersions } from '@looker/sdk-codegen'
 import { useLocation } from 'react-router-dom'
-import { appPath, getEnvAdaptor } from '..'
+import { appPath, getEnvAdaptor } from '../../..'
+import { getVersions, validateUrl } from '../utils'
+import { useOAuthFormActions, useOAuthFormState } from '../state/slice'
+import { CollapserCard } from './CollapserCard'
 import { ConfigHeading } from './ConfigHeading'
-import { CollapserCard } from './Collapser'
-import { getVersions, validateUrl } from './utils'
-import type { ConfigValues } from './utils'
 
 export const readyToLogin =
   'OAuth is configured but your browser session is not authenticated. Click Login.'
-
-const POSITIVE: MessageBarIntent = 'positive'
-
-interface IFieldValues {
-  baseUrl: string
-  fetchedUrl: string
-  webUrl: string
-  fetchIntent: MessageBarIntent
-  fetchResult: string
-}
-
-const defaultFieldValues: IFieldValues = {
-  baseUrl: '',
-  fetchedUrl: '',
-  webUrl: '',
-  fetchResult: '',
-  fetchIntent: POSITIVE,
-}
 
 const EmptyConfig = { base_url: '', looker_url: '' }
 
@@ -84,7 +66,7 @@ interface ConfigFormProps {
   clientLabel: string
 }
 
-export const OAuthConfigForm = ({
+export const OAuthForm = ({
   setHasConfig,
   clientId,
   clientLabel,
@@ -92,81 +74,43 @@ export const OAuthConfigForm = ({
 }: ConfigFormProps) => {
   const location = useLocation()
   const redirect_uri = appPath(location, '/oauth')
-  const BASE_URL = 'baseUrl'
-  const WEB_URL = 'webUrl'
-  const FETCH_INTENT = 'fetchIntent'
-  const FETCH_RESULT = 'fetchResult'
-  const CRITICAL: MessageBarIntent = 'critical'
   const appConfig = `// client_guid=${clientId}
-{
-  "redirect_uri": "${redirect_uri}",
-  "display_name": "CORS ${clientLabel}",
-  "description": "Looker ${clientLabel} using CORS",
-  "enabled": true
-}
-`
+ {
+   "redirect_uri": "${redirect_uri}",
+   "display_name": "CORS ${clientLabel}",
+   "description": "Looker ${clientLabel} using CORS",
+   "enabled": true
+ }
+ `
   const adaptor = getEnvAdaptor()
   const sdk = adaptor.sdk
-  // See https://codesandbox.io/s/youthful-surf-0g27j?file=/src/index.tsx for a prototype from Luke
-  // TODO see about useReducer to clean this up a bit more
 
-  const getConfig = () => {
-    // TODO: This is temporary until config settings are available in redux.
-    // get configuration from storage, or default it
-    const data = localStorage.getItem(configKey)
-    const result = data ? JSON.parse(data) : EmptyConfig
-    return result
-  }
-
-  const config = getConfig()
-  const [fields, setFields] = useState<IFieldValues>(defaultFieldValues)
-  const [saved, setSaved] = useState<ConfigValues>(config)
-
-  const updateFields = (
-    nameOrValues: string | Partial<IFieldValues>,
-    value = ''
-  ) => {
-    if (typeof nameOrValues === 'string') {
-      setFields((previousFields) => {
-        return { ...previousFields, ...{ [nameOrValues]: value } }
-      })
-    } else {
-      setFields((previousFields) => {
-        return { ...previousFields, ...nameOrValues }
-      })
-    }
-  }
-
-  useEffect(() => {
-    const data = getConfig()
-    const { base_url, looker_url } = data
-    setSaved(data)
-    updateFields({
-      [BASE_URL]: base_url,
-      [WEB_URL]: looker_url,
-      [FETCH_INTENT]:
-        base_url !== '' && looker_url !== '' ? POSITIVE : CRITICAL,
-    })
-  }, [])
-
-  const [validationMessages, setValidationMessages] =
-    useState<ValidationMessages>({})
-
-  const updateMessage = (intent: MessageBarIntent, message: string) => {
-    updateFields({ [FETCH_RESULT]: message, [FETCH_INTENT]: intent })
-  }
+  const {
+    setApiServerUrlAction,
+    updateApiServerUrlAction,
+    setFetchedUrlAction,
+    setWebUrlAction,
+    clearFormAction,
+    updateMessageBarAction,
+    verifyErrorAction,
+    saveNewConfigAction,
+  } = useOAuthFormActions()
+  const {
+    apiServerUrl,
+    fetchedUrl,
+    webUrl,
+    messageBar,
+    validationMessages,
+    savedConfig,
+  } = useOAuthFormState()
 
   const isConfigured = () => {
     return (
-      saved !== EmptyConfig &&
-      fields[BASE_URL] === saved.base_url &&
-      fields[WEB_URL] === saved.looker_url
+      apiServerUrl === savedConfig.base_url &&
+      webUrl === savedConfig.looker_url &&
+      apiServerUrl !== '' &&
+      webUrl !== ''
     )
-  }
-
-  const fetchError = (message: string) => {
-    updateFields(WEB_URL, '')
-    updateMessage(CRITICAL, message)
   }
 
   const saveConfig = (baseUrl: string, webUrl: string) => {
@@ -176,97 +120,94 @@ export const OAuthConfigForm = ({
       client_id: clientId,
       redirect_uri,
     }
-    updateFields({
-      [BASE_URL]: baseUrl,
-      [WEB_URL]: webUrl,
-    })
-    // TODO: replace when redux is introduced
     localStorage.setItem(configKey, JSON.stringify(data))
+    saveNewConfigAction(data)
     if (setHasConfig) setHasConfig(true)
-    setSaved(data)
-    updateMessage(POSITIVE, `Saved ${webUrl} as OAuth server`)
   }
 
-  const updateForm = async (_e: BaseSyntheticEvent, save: boolean) => {
-    updateMessage('inform', '')
-    const versionsUrl = `${fields.baseUrl}/versions`
+  const verifyUrl = async (): Promise<ILookerVersions | undefined> => {
+    updateMessageBarAction({ intent: messageBar.intent, text: '' })
+    const versionsUrl = `${apiServerUrl}/versions`
     try {
-      updateFields({
-        fetchedUrl: fields.baseUrl,
-      })
-      const { web_server_url: webUrl, api_server_url: baseUrl } =
-        (await getVersions(versionsUrl)) as ILookerVersions
-      updateMessage(POSITIVE, 'Configuration is valid')
-      updateFields({
-        [BASE_URL]: baseUrl,
-        [WEB_URL]: webUrl,
-      })
-      if (save) {
-        saveConfig(baseUrl, webUrl)
+      setFetchedUrlAction(apiServerUrl)
+
+      const versions = await getVersions(versionsUrl)
+
+      if (versions) {
+        updateMessageBarAction({
+          intent: 'positive',
+          text: `Configuration is valid`,
+        })
+        setWebUrlAction(versions.web_server_url)
       }
+
+      return versions
     } catch (e: any) {
-      fetchError(e.message)
+      verifyErrorAction(e.message)
+      return undefined
     }
   }
 
-  const handleSave = async (e: BaseSyntheticEvent) => {
-    await updateForm(e, true)
+  const handleSave = async () => {
+    const versions = await verifyUrl()
+    if (versions) {
+      saveConfig(versions.api_server_url, versions.web_server_url)
+    }
   }
 
-  const handleVerify = async (e: BaseSyntheticEvent) => {
-    await updateForm(e, false)
+  const handleVerify = () => {
+    verifyUrl()
   }
 
   const handleClear = async (_e: BaseSyntheticEvent) => {
-    // TODO: replace when redux is introduced to run it
     localStorage.removeItem(configKey)
-    updateFields({
-      [BASE_URL]: '',
-      [WEB_URL]: '',
-      [FETCH_INTENT]: CRITICAL,
-      [FETCH_RESULT]: '',
-    })
-    setSaved(EmptyConfig)
+    clearFormAction()
+
     if (setHasConfig) setHasConfig(false)
     if (isAuthenticated()) {
-      updateMessage('warn', 'Please reload the browser page to log out')
+      updateMessageBarAction({
+        intent: 'warn',
+        text: `Please reload the browser page to log out`,
+      })
     }
   }
 
   const handleUrlChange = (event: FormEvent<HTMLInputElement>) => {
     const name = event.currentTarget.name
+    const value = event.currentTarget.value
+    let newApiServerUrl = value
 
     const newValidationMessages = { ...validationMessages }
 
-    const url = validateUrl(event.currentTarget.value)
+    const url = validateUrl(value)
     if (url) {
       delete newValidationMessages[name]
       // Update URL if it's been cleaned up
-      event.currentTarget.value = url
+      newApiServerUrl = url
     } else {
       newValidationMessages[name] = {
-        message: `'${event.currentTarget.value}' is not a valid url`,
+        message: `'${value}' is not a valid url`,
         type: 'error',
       }
     }
-    updateFields({
-      [event.currentTarget.name]: event.currentTarget.value,
-      [WEB_URL]: '',
-    })
 
-    setValidationMessages(newValidationMessages)
+    updateApiServerUrlAction({
+      apiServerUrl: newApiServerUrl,
+      webUrl: '',
+      validationMessages: newValidationMessages as ValidationMessages,
+    })
   }
 
   const isAuthenticated = () => sdk.authSession.isAuthenticated()
 
   const verifyButtonDisabled =
-    fields.baseUrl.trim().length === 0 ||
+    apiServerUrl.trim().length === 0 ||
     Object.keys(validationMessages).length > 0
 
   const saveButtonDisabled =
-    verifyButtonDisabled || fields.webUrl.trim().length === 0 || isConfigured()
+    verifyButtonDisabled || webUrl.trim().length === 0 || isConfigured()
 
-  const clearButtonDisabled = fields.baseUrl.trim().length === 0
+  const clearButtonDisabled = apiServerUrl.trim().length === 0
 
   const loginButtonDisabled =
     verifyButtonDisabled || !isConfigured() || isAuthenticated()
@@ -277,6 +218,17 @@ export const OAuthConfigForm = ({
     await adaptor.login()
   }
 
+  useEffect(() => {
+    // Get config values from localstorage
+    const data = localStorage.getItem(configKey)
+    const result = data ? JSON.parse(data) : EmptyConfig
+    if (result.base_url && result.looker_url) {
+      const { base_url, looker_url } = result
+      setApiServerUrlAction(base_url)
+      setWebUrlAction(looker_url)
+    }
+  }, [])
+
   return (
     <SpaceVertical gap="u2">
       <ConfigHeading>Looker OAuth Configuration</ConfigHeading>
@@ -285,11 +237,13 @@ export const OAuthConfigForm = ({
         authenticate into your Looker Instance
       </Span>
       <MessageBar
-        intent={fields[FETCH_INTENT]}
-        onPrimaryClick={() => updateMessage(fields[FETCH_INTENT], '')}
-        visible={fields[FETCH_RESULT] !== ''}
+        intent={messageBar.intent}
+        onPrimaryClick={() =>
+          updateMessageBarAction({ intent: messageBar.intent, text: '' })
+        }
+        visible={messageBar.text !== ''}
       >
-        {fields[FETCH_RESULT]}
+        {messageBar.text}
       </MessageBar>
       <CollapserCard
         heading="1. Supply API Server URL"
@@ -304,23 +258,23 @@ export const OAuthConfigForm = ({
                 required
                 label="API server URL"
                 placeholder="typically https://myserver.looker.com:19999"
-                name={BASE_URL}
-                value={fields[BASE_URL]}
+                name="baseUrl"
+                value={apiServerUrl}
                 onChange={handleUrlChange}
               />
               <FieldText
                 label="OAuth server URL"
                 placeholder="Click 'Verify' to retrieve"
-                name={WEB_URL}
-                value={fields[WEB_URL]}
+                name="webUrl"
+                value={webUrl}
                 disabled={true}
               />
             </Fieldset>
           </Form>
-          {fields.fetchedUrl && (
+          {fetchedUrl && (
             <>
               <Paragraph fontSize="small">
-                On {fields.fetchedUrl}, enable {clientLabel} as a{' '}
+                On {fetchedUrl}, enable {clientLabel} as a{' '}
                 <Link
                   href="https://github.com/looker-open-source/sdk-codegen/blob/main/docs/cors.md#reference-implementation"
                   target="_blank"
@@ -328,13 +282,13 @@ export const OAuthConfigForm = ({
                   Looker OAuth client
                 </Link>{' '}
                 by adding "{(window as any).location.origin}" to the{' '}
-                <Link href={`${fields.webUrl}/admin/embed`} target="_blank">
+                <Link href={`${webUrl}/admin/embed`} target="_blank">
                   Embedded Domain Allowlist
                 </Link>
                 . If API Explorer is also installed, the configuration below can
                 be used to{' '}
                 <Link
-                  href={`${fields.fetchedUrl}/extensions/marketplace_extension_api_explorer::api-explorer/4.0/methods/Auth/register_oauth_client_app`}
+                  href={`${fetchedUrl}/extensions/marketplace_extension_api_explorer::api-explorer/4.0/methods/Auth/register_oauth_client_app`}
                   target="_blank"
                 >
                   register this API Explorer instance
@@ -353,7 +307,7 @@ export const OAuthConfigForm = ({
                 Clear
               </ButtonTransparent>
             </Tooltip>
-            <Tooltip content={`Verify ${fields[BASE_URL]}`}>
+            <Tooltip content={`Verify ${apiServerUrl}`}>
               <ButtonTransparent
                 disabled={verifyButtonDisabled}
                 onClick={handleVerify}
@@ -393,7 +347,7 @@ export const OAuthConfigForm = ({
               You will be able to login after you Verify your API Server URL
             </Span>
           )}
-          <Tooltip content={`Login to ${fields[WEB_URL]} using OAuth`}>
+          <Tooltip content={`Login to ${webUrl} using OAuth`}>
             <Button onClick={handleLogin} disabled={loginButtonDisabled}>
               Login
             </Button>