+++ packages/sdk/src/4.0/models.ts
@@ -2721,7 +2721,7 @@ export interface IDBConnection {
    */
   query_timezone?: string | null
   /**
-   * Scheme name
+   * Schema name
    */
   schema?: string | null
   /**
@@ -2905,7 +2905,7 @@ export interface IDBConnectionOverride {
    */
   database?: string | null
   /**
-   * Scheme name
+   * Schema name
    */
   schema?: string | null
   /**
@@ -7565,6 +7565,10 @@ export interface IRequestContentThumbnail {
    */
   reload?: string | null
   /**
+   * Light or dark background. Default is "light"
+   */
+  theme?: string | null
+  /**
    * A value of png produces a thumbnail in PNG format instead of SVG (default)
    */
   format?: string | null
@@ -7833,11 +7837,11 @@ export interface IRequestGraphDerivedTablesForView {
  */
 export interface IRequestLogin {
   /**
-   * client_id part of API3 Key.
+   * client_id part of API Key.
    */
   client_id?: string | null
   /**
-   * client_secret part of API3 Key.
+   * client_secret part of API Key.
    */
   client_secret?: string | null
 }
@@ -10729,7 +10733,7 @@ export interface IThemeSettings {
    */
   font_source?: string | null
   /**
-   * Info button color
+   * (DEPRECATED) Info button color
    */
   info_button_color?: string
   /**
@@ -10765,7 +10769,7 @@ export interface IThemeSettings {
    */
   title_color?: string
   /**
-   * Warning button color
+   * (DEPRECATED) Warning button color
    */
   warn_button_color?: string
   /**
@@ -10800,6 +10804,14 @@ export interface IThemeSettings {
    * Toggle to center the dashboard title. Defaults to false.
    */
   center_dashboard_title?: boolean
+  /**
+   * Dashboard title font size.
+   */
+  dashboard_title_font_size?: string
+  /**
+   * Default box shadow.
+   */
+  box_shadow?: string
 }
 
 export interface ITimezone {
@@ -10861,7 +10873,7 @@ export interface IUser {
    */
   avatar_url_without_sizing?: string | null
   /**
-   * API 3 credentials (read-only)
+   * API credentials (read-only)
    */
   credentials_api3?: ICredentialsApi3[] | null
   credentials_email?: ICredentialsEmail
@@ -12141,7 +12153,7 @@ export interface IWriteDBConnection {
    */
   query_timezone?: string | null
   /**
-   * Scheme name
+   * Schema name
    */
   schema?: string | null
   /**
@@ -12281,7 +12293,7 @@ export interface IWriteDBConnectionOverride {
    */
   database?: string | null
   /**
-   * Scheme name
+   * Schema name
    */
   schema?: string | null
   /**