+++ packages/extension-api-explorer/CHANGELOG.md
@@ -22,6 +22,17 @@
     * @looker/run-it bumped from ^0.9.44 to ^0.9.45
     * @looker/sdk bumped from ^23.2.0 to ^23.4.0
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @looker/api-explorer bumped from ^0.9.45 to ^0.9.46
+    * @looker/extension-sdk bumped from ^23.4.0 to ^23.6.0
+    * @looker/extension-sdk-react bumped from ^23.4.0 to ^23.6.0
+    * @looker/extension-utils bumped from ^0.1.22 to ^0.1.23
+    * @looker/run-it bumped from ^0.9.45 to ^0.9.46
+    * @looker/sdk bumped from ^23.4.0 to ^23.6.0
+
 ## [22.21.0](https://github.com/looker-open-source/sdk-codegen/compare/extension-api-explorer-v22.20.1...extension-api-explorer-v22.21.0) (2023-02-22)